changed output to sigmoid and changed loss to binary xentropy

+++ model_3.py
@@ -60,7 +60,7 @@ def main():
 
     # cost = tf.reduce_mean(tf.nn.sparse_softmax_cross_entropy_with_logits(logits, y))
     one = tf.constant(1, dtype=tf.int64)
-    cost = tf.reduce_mean(-y * tf.to_int64(tf.log(logits)) - (one - y) * tf.to_int64(tf.log(logits)))  # binary xentropy
+    cost = tf.reduce_mean(tf.cast(-y * tf.to_int64(tf.log(logits)) - (one - y) * tf.to_int64(tf.log(logits)), dtype=tf.float32)) # binary xentropy
     train_prediction = tf.nn.softmax(logits)
     optimizer = tf.train.AdamOptimizer(learning_rate=params['lr']).minimize(cost)
     tf.scalar_summary("cost", cost)
